корректировка для защиты
