--- js/data.js
@@ -1,100 +1,0 @@
-import {getRandomFloat, getRandomInteger} from './util-functions.js';
-
-const IMG_ADDRESS = 'img/avatars/user';
-
-const TITLES = [
-  'У нас вы забудете обо всем на свете!',
-  'Отдохнете - мама родная не узнает!',
-  'Одиссей, тебе сюда от жены от детей!',
-  'Приезжай, торопись! Отдыхай - не скупись!',
-  'Отдохнете лучше, чем сын маминой подруги!',
-  'Клюет лучше, чем на черных камнях!',
-  'Приехал, заплатил, отдохнул!',
-  'Сколько волка не корми, а он все равно у нас отдыхать любит!',
-  'Семь раз отмерь, а отдохни у нас!',
-  'Хоромы царские, перины мягкие!',
-];
-
-export const TYPES = {
-  palace : 'Дворец',
-  flat: 'Квартира',
-  house: 'Дом',
-  bungalow: 'Бунгало',
-};
-
-export const CHECKIN_TIMES = ['12:00', '13:00', '14:00' ];
-
-export const CHECKOUT_TIMES = ['12:00', '13:00', '14:00' ];
-
-const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner' ];
-
-const DESCRIPTION = [
-  'Тепло, сухо, уютно',
-  'Красиво, уютно, чисто',
-  'Есть всё',
-  'Светло, тепло, уютно',
-  'Много пространства для отдыха',
-  'Есть вся необходимая техника',
-  'Скромно, но со вкусом',
-  'Окна с видом на море',
-  'Тихие соседи и их собаки',
-  'Уютное, теплое помещение. Почти как у Гарри',
-];
-
-const PHOTOS = [
-  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
-];
-
-const getRandomSentence = (elements) => {
-  return elements[getRandomInteger(0, elements.length - 1)];
-};
-
-const getRandomArray = (array) => {
-  return array.slice(0, getRandomInteger(1, array.length));
-};
-
-const getRandomAvatar = (imgAddressArray) => {
-  return {
-    avatar: imgAddressArray + `${getRandomInteger(1, 8)}`.padStart(2, '0') + '.png',
-  };
-};
-
-const getRandomObjectLocation = () => {
-  return {
-    x: getRandomFloat(35.65000, 35.70000, 5),
-    y: getRandomFloat(139.70000, 139.80000, 5),
-  }
-};
-
-
-const getRandomOffer = () => {
-  return {
-    title: getRandomSentence(TITLES),
-    address: Object.values(getRandomObjectLocation()).join(', '),
-    price: getRandomInteger(100, 1000),
-    type: getRandomSentence(Object.keys(TYPES)),
-    rooms: getRandomInteger(1, 6),
-    get guests() {
-      return this.rooms + 2;
-    },
-    set guests(value) {
-      this.guests = value;
-    },
-    checkin: getRandomSentence(CHECKIN_TIMES),
-    checkout: getRandomSentence(CHECKOUT_TIMES),
-    features: getRandomArray(FEATURES),
-    description: getRandomSentence(DESCRIPTION),
-    photos: getRandomArray(PHOTOS),
-  }
-};
-
-
-export const getNewBookingObject = () => {
-  return {
-    author: getRandomAvatar(IMG_ADDRESS),
-    offer: getRandomOffer(),
-    objectLocation: getRandomObjectLocation(),
-  };
-};
